STAC-22973: drop kafka-go usage, simplify franz-go usage
